+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -343,11 +343,13 @@ public class PostHog private constructor(
             props.putAll(it)
         }
 
+        val isSessionReplayFlagActive = isSessionReplayFlagActive()
+
         PostHogSessionManager.getActiveSessionId()?.let { sessionId ->
             val tempSessionId = sessionId.toString()
             props["\$session_id"] = tempSessionId
             // only Session replay needs $window_id
-            if (!appendSharedProps && config?.sessionReplay == true) {
+            if (!appendSharedProps && isSessionReplayFlagActive) {
                 // Session replay requires $window_id, so we set as the same as $session_id.
                 // the backend might fallback to $session_id if $window_id is not present next.
                 props["\$window_id"] = tempSessionId
@@ -361,7 +363,7 @@ public class PostHog private constructor(
         // only Session replay needs distinct_id also in the props
         // remove after https://github.com/PostHog/posthog/pull/18954 gets merged
         val propDistinctId = props["distinct_id"] as? String
-        if (!appendSharedProps && config?.sessionReplay == true && propDistinctId.isNullOrBlank()) {
+        if (!appendSharedProps && isSessionReplayFlagActive && propDistinctId.isNullOrBlank()) {
             // distinctId is already validated hence not empty or blank
             props["distinct_id"] = distinctId
         }
@@ -838,6 +840,17 @@ public class PostHog private constructor(
         return PostHogSessionManager.isSessionActive()
     }
 
+    // this is used in cases where we know the session is already active
+    // so we spare another locker
+    private fun isSessionReplayFlagActive(): Boolean {
+        return config?.sessionReplay == true && featureFlags?.isSessionReplayFlagActive() == true
+    }
+
+    override fun isSessionReplayActive(): Boolean {
+        // not checking isEnabled() here because isSessionActive already does that anyway
+        return isSessionReplayFlagActive() && isSessionActive()
+    }
+
     override fun getSessionId(): UUID? {
         if (!isEnabled()) {
             return null
@@ -1022,6 +1035,10 @@ public class PostHog private constructor(
 
         override fun isSessionActive(): Boolean {
             return shared.isSessionActive()
+        }
+
+        override fun isSessionReplayActive(): Boolean {
+            return shared.isSessionReplayActive()
         }
 
         override fun getSessionId(): UUID? {